feat: 장소 페이지 버튼 수정
