+++ src/components/poi/poi-images.tsx
@@ -1,15 +1,28 @@
-const POIImages = () => {
+import { type POI } from '@/types/pois';
+import { AspectRatio } from '@radix-ui/react-aspect-ratio';
+
+type POIImagesProps = {
+  images: POI['imgs_on_poi'];
+};
+
+const POIImages = (props: POIImagesProps) => {
+  const { images } = props;
   return (
-    <section className='flex flex-col gap-2 overflow-x-auto p-4'>
-      <h2 className='text-16M'>사진 · 5장</h2>
-      <ul className='flex gap-2 overflow-auto'>
-        <li className='h-28 w-20 flex-none rounded-sm bg-gray-3'></li>
-        <li className='h-28 w-20 flex-none rounded-sm bg-gray-3'></li>
-        <li className='h-28 w-20 flex-none rounded-sm bg-gray-3'></li>
-        <li className='h-28 w-20 flex-none rounded-sm bg-gray-3'></li>
-        <li className='h-28 w-20 flex-none rounded-sm bg-gray-3'></li>
-      </ul>
-    </section>
+    <ul className='flex flex-nowrap gap-2 overflow-hidden overflow-x-auto px-4'>
+      {[...images, ...images].map((image, idx) => (
+        <li className='h-full w-20 flex-none' key={`${image.img_id}-${idx}`}>
+          <AspectRatio ratio={12 / 17} className='overflow-hidden rounded-sm'>
+            <img
+              className='object-cover'
+              src={
+                'https://images.unsplash.com/photo-1535025183041-0991a977e25b?w=300&dpr=2&q=80'
+              }
+              alt='Landscape photograph by Tobias Tullius'
+            />
+          </AspectRatio>
+        </li>
+      ))}
+    </ul>
   );
 };
